Fix warning

+++ test/test_blocked_elements.py
@@ -106,7 +106,7 @@ def test_tensor_element(compile_args):
     assert ufc_dofmap.num_entity_dofs[2] == 0
     assert ufc_dofmap.num_entity_dofs[3] == 0
     for v in range(3):
-        vals = np.zeros(1, dtype=np.int)
+        vals = np.zeros(1, dtype=np.int32)
         vals_ptr = module.ffi.cast("int *", module.ffi.from_buffer(vals))
         ufc_dofmap.tabulate_entity_dofs(vals_ptr, 0, v)
         assert vals[0] == v
